Enable fetching of db content into view.
Requiring angular-resource gives us a custom service that represents a RESTful client. We will use this to communicate via JSON to our rails routes for CRUD implementation. Note that every action in our EntriesController uses respond_to :json and respond_with.

app = angular.module("Raffler", ["ngResource"])

At the top of our angular controller, we give our app a name and pass it dependencies, namely the "ngResource" file that we now have access to.

<html ng-app="Raffler">

In order for everything to work, we have to pass this app name into a DOM node that will contain our app logic.

@RaffleCtrl = ($scope, $resource) ->
  Entry = $resource("/entries/:id", {id: "@id"}, {update: {method: "PUT"}})
  $scope.entries = Entry.query()

$resource is a function we can call to return an object that allows us to communicate over our REST API. The first paramater is the url we want it to use. The second is the parameters we want to pass - in this case, the current object'd id, if it has one. (If not the url for Entry to use will be /entries.) The third argument is optional actions you want it to use; most of our actions were accounted for by default, but we do have to define an update action, and tell it to use the PUT method.

http://docs.angularjs.org/api/ngResource.$resource
See the resource docs; there are action keywords which correspond to HTTP methods. Entry.query() corresponds to a GET request on a collection ( .get() just refers to a single resource. )